requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
SQLAlchemy==2.0.27
python-dotenv==1.0.0
pandas==2.2.1
fastapi==0.104.1
uvicorn==0.24.0
langchain-core>=0.2.27
langchain>=0.1.0
langchain-community>=0.0.10
openai==1.3.5
faiss-cpu==1.7.4
pydantic==2.5.2
langgraph==0.2.3
llama-cpp-python==0.2.56
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99
accelerate==0.25.0
tqdm==4.66.2
requests==2.31.0
sentence-transformers==2.5.1
numpy>=1.24.0
scipy>=1.10.0 